refactor(qbtc): address CodeRabbit review on #4092
- Extract cosmosStakedFiat to a shared util and reuse it in the QBTC banner
  and the DeFi portfolio rollup (removes duplicated staked-fiat math).
- Drop the `as QbtcDappMessage[]` assertion in the staking keysign builder;
  map the cosmjs EncodeObject[] to the QBTC dApp message shape instead.
- Add JSDoc to the exported staking query hooks (delegations / rewards /
  unbondings), the validator/delegation pickers, and StakeOverview.
